doc: add comments and documentation
aslo add docker file

+++ benchmark/random_write.rs
@@ -38,17 +38,17 @@ pub fn gen_rand_string(n: usize) -> String {
         .chars()
         .collect();
     let mut b = vec!['x'; n];
-    let mut i = n - 1;
+    let mut i = (n - 1) as i64;
     let mut cache = rand::random::<u64>();
     let mut remain = LETTER_IDX_MAX;
-    while i as i64 >= 0 {
+    while i >= 0 {
         if remain == 0 {
             cache = rand::random::<u64>();
             remain = LETTER_IDX_MAX;
         }
         let idx: u32 = (cache & LETTER_IDX_MASK) as u32;
         if idx < letters.len() as u32 {
-            b[i] = letters[idx as usize];
+            b[i as usize] = letters[idx as usize];
             i -= 1;
         }
         cache >>= LETTER_IDX_BITS;
